refactor: delete logging package
