+++ src/main/kotlin/org/rooftop/netx/redis/RedisTransactionConfigurer.kt
@@ -3,7 +3,7 @@ package org.rooftop.netx.redis
 import org.rooftop.netx.api.TransactionManager
 import org.rooftop.netx.engine.info
 import org.rooftop.netx.engine.logger
-import org.rooftop.netx.logging.LoggerFactory
+import org.rooftop.netx.engine.LoggerFactory
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
 import org.springframework.context.ApplicationContext
